refactor(auction): widen timestamp to uint64, rename event, add tests

+++ contracts/src/libraries/Types.sol
@@ -223,7 +223,7 @@ library Types {
         Types.DutchAuctionParams dutchAuctionParams;
         uint256 depositAmount;
         uint256 startBlock;
-        uint32 startTimestamp;
+        uint64 startTimestamp;
         DutchAuctionState state;
     }
